refactor: remove IssuesPage and related routes from the dashboard and move it to flow runs
